Slaatt sammen core, local og domain
