+++ .github/workflows/deploy.testnorge-sam.yml
@@ -2,10 +2,11 @@ name: Deploy testnorge-sam
 
 on:
   push:
-    branches: master
+    branches:
+      - master
     paths:
       - apps/testnorge-sam/**
-
+  workflow_dispatch:
 
 env:
   IMAGE: docker.pkg.github.com/${{ github.repository }}/sam:${{ github.sha }}
