--- apps/testnorge-sam/sam-core/build.gradle
@@ -1,52 +1,0 @@
-plugins {
-    id 'no.nav.registre.testnorge.java-conventions'
-    id 'org.springframework.boot' version "${springBootVersion}"
-    id 'io.spring.dependency-management' version "${springDependencyManagement}"
-}
-
-dependencyManagement {
-    imports {
-        mavenBom  "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
-}
-
-bootJar {
-    archiveFileName = "app.jar"
-}
-
-dependencies {
-    implementation 'no.nav.testnav.libs:servlet-core'
-    implementation project(':sam-domain')
-
-    implementation 'org.springframework.cloud:spring-cloud-starter-vault-config'
-    implementation 'org.springframework.cloud:spring-cloud-starter-sleuth'
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter'
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.springframework.cloud:spring-cloud-starter-bootstrap' // TODO remove legacy bootstrap config
-    implementation 'org.apache.httpcomponents:httpclient:4.5.6'
-    implementation 'org.aspectj:aspectjweaver:1.9.6'
-    implementation 'io.micrometer:micrometer-registry-prometheus:1.5.5'
-
-    implementation 'io.springfox:springfox-swagger2:2.9.2'
-    implementation 'io.springfox:springfox-swagger-ui:2.9.2'
-    implementation "net.logstash.logback:logstash-logback-encoder:${logstashLogbackEncoder}"
-
-
-    implementation 'javax.xml.bind:jaxb-api:2.3.1'
-    implementation 'no.nav.registre:testnorge-consumers:1.12.0-RELEASE'
-
-    runtimeOnly 'org.hibernate:hibernate-entitymanager:5.2.10.Final'
-    runtimeOnly 'mysql:mysql-connector-java:6.0.6'
-    runtimeOnly 'com.oracle.database.jdbc:ojdbc8:19.7.0.0'
-
-    testImplementation 'junit:junit:4.12' //TODO upgrade to JUnit5
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    testImplementation 'org.springframework.cloud:spring-cloud-contract-wiremock'
-    testImplementation 'com.h2database:h2:1.4.195'
-}
-
-description = 'sam-core'
