--- apps/testnorge-sam/sam-domain/build.gradle
@@ -1,10 +1,0 @@
-plugins {
-    id 'no.nav.registre.testnorge.java-conventions'
-}
-
-dependencies {
-
-    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.9.0'
-}
-
-description = 'sam-domain'
