--- apps/testnorge-sam/sam-domain/src/main/java/no/nav/registre/sam/IdentMedData.java
@@ -1,19 +1,0 @@
-package no.nav.registre.sam;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.util.List;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Getter
-@Setter
-public class IdentMedData {
-    private String id;
-    private List<SyntetisertSamordningsmelding> data;
-}
