--- apps/testnorge-sam/sam-local/build.gradle
@@ -1,25 +1,0 @@
-plugins {
-    id 'no.nav.registre.testnorge.java-conventions'
-    id 'org.springframework.boot' version "${springBootVersion}"
-    id 'io.spring.dependency-management' version "${springDependencyManagement}"
-}
-bootJar {
-    archiveFileName = "app.jar"
-}
-
-dependencyManagement {
-    imports {
-        mavenBom  "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
-    }
-}
-
-
-dependencies {
-    implementation project(':sam-core')
-    implementation 'org.springframework.boot:spring-boot-autoconfigure'
-    implementation 'org.hibernate.validator:hibernate-validator'
-    testImplementation 'junit:junit:4.12' //TODO upgrade to JUnit5
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-}
-
-description = 'sam-local'
