--- apps/testnorge-sam/sam-local/src/main/java/no/nav/registre/sam/LocalApplicationStarter.java
@@ -1,15 +1,0 @@
-package no.nav.registre.sam;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.builder.SpringApplicationBuilder;
-
-@SpringBootApplication
-public class LocalApplicationStarter {
-
-    public static void main(String[] args) {
-        new SpringApplicationBuilder()
-                .sources(LocalApplicationStarter.class)
-                .profiles("local")
-                .run(args);
-    }
-}
