+++ apps/testnorge-sam/src/main/java/no/nav/registre/sam/ApplicationStarter.java
