+++ apps/testnorge-sam/src/main/java/no/nav/registre/sam/database/TSamVedtakRepository.java
