+++ apps/testnorge-sam/src/main/java/no/nav/registre/sam/domain/IdentMedData.java
@@ -1,0 +1,19 @@
+package no.nav.registre.sam.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.List;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Getter
+@Setter
+public class IdentMedData {
+    private String id;
+    private List<SyntetisertSamordningsmelding> data;
+}
