+++ apps/testnorge-sam/src/main/java/no/nav/registre/sam/multitenancy/TenantContext.java
