+++ apps/testnorge-sam/src/main/resources/application-local.properties
