+++ apps/testnorge-sam/src/main/resources/application-prod.properties
