+++ apps/testnorge-sam/src/main/resources/application.properties
