+++ apps/testnorge-sam/src/main/resources/bootstrap.yml
