+++ apps/testnorge-sam/src/main/resources/logback-spring.xml
