+++ apps/testnorge-sam/src/test/java/no/nav/registre/sam/consumer/rs/SamSyntetisererenConsumerTest.java
