+++ apps/testnorge-sam/src/test/java/no/nav/registre/sam/testutils/ResourceUtils.java
