+++ apps/testnorge-sam/src/test/resources/application-test.properties
