+++ settings.gradle
@@ -112,7 +112,6 @@ include(':sigrun-local')
 include(':testnorge-nav-endringsmeldinger')
 include(':testnorge-frikort')
 include(':testnorge-aaregstub')
-include(':sam-domain')
 include(':medl-domain')
 include(':testnorge-sigrun')
 include(':testnorge-token-provider')
@@ -126,9 +125,7 @@ include(':testnorge-statisk-data-forvalter')
 include(':nav-endringsmeldinger-local')
 include(':tp-local')
 include(':testnorge-populasjoner')
-include(':sam-core')
 include(':tilbakemelding-api')
-include(':sam-local')
 include(':mn-organisasjon-api')
 include(':organisasjon-service')
 include(':organisasjon-service')
@@ -163,7 +160,6 @@ project(":sigrun-local").projectDir = file('apps/testnorge-sigrun/sigrun-local')
 project(":testnorge-nav-endringsmeldinger").projectDir = file('apps/testnorge-nav-endringsmeldinger')
 project(":testnorge-frikort").projectDir = file('apps/testnorge-frikort')
 project(":testnorge-aaregstub").projectDir = file('apps/testnorge-aaregstub')
-project(":sam-domain").projectDir = file('apps/testnorge-sam/sam-domain')
 project(":medl-domain").projectDir = file('apps/testnorge-medl/medl-domain')
 project(":testnorge-sigrun").projectDir = file('apps/testnorge-sigrun')
 project(":testnorge-token-provider").projectDir = file('apps/testnorge-token-provider')
@@ -177,9 +173,7 @@ project(":testnorge-statisk-data-forvalter").projectDir = file('apps/testnorge-s
 project(":nav-endringsmeldinger-local").projectDir = file('apps/testnorge-nav-endringsmeldinger/nav-endringsmeldinger-local')
 project(":tp-local").projectDir = file('apps/testnorge-tp/tp-local')
 project(":testnorge-populasjoner").projectDir = file('apps/testnorge-populasjoner')
-project(":sam-core").projectDir = file('apps/testnorge-sam/sam-core')
 project(":tilbakemelding-api").projectDir = file('apps/tilbakemelding-api')
-project(":sam-local").projectDir = file('apps/testnorge-sam/sam-local')
 project(":mn-organisasjon-api").projectDir = file('apps/mn-organisasjon-api')
 project(":organisasjon-service").projectDir = file('apps/organisasjon-service')
 project(":organisasjon-service").projectDir = file('apps/organisasjon-service')
